Split cli into automatic and manual modules.
