+++ src/cli.ts
@@ -13,37 +13,22 @@ require('source-map-support').install();
 
 import * as fsExtra from 'fs-extra';
 import * as path from 'path';
-import * as webdriver from 'selenium-webdriver';
 
 import commandLineUsage = require('command-line-usage');
-import ProgressBar = require('progress');
-import ansi = require('ansi-escape-sequences');
 
 import {optDefs, parseFlags} from './flags';
-import {makeSession} from './session';
-import {browserSignature, fcpBrowsers, makeDriver, openAndSwitchToNewTab, pollForGlobalResult, pollForFirstContentfulPaint} from './browser';
-import {BenchmarkResult, BenchmarkSpec} from './types';
+import {fcpBrowsers} from './browser';
+import {BenchmarkSpec} from './types';
 import {Server} from './server';
-import {formatCsv} from './csv';
-import {Horizons, ResultStats, ResultStatsWithDifferences, horizonsResolved, summaryStats, computeDifferences} from './stats';
+import {Horizons, ResultStatsWithDifferences} from './stats';
 import {specsFromOpts} from './specs';
-import {AutomaticResults, verticalTermResultTable, horizontalTermResultTable, verticalHtmlResultTable, horizontalHtmlResultTable, automaticResultTable, spinner, benchmarkOneLiner} from './format';
 import {prepareVersionDirectory, makeServerPlans} from './versions';
 import {Config} from './config';
 import * as defaults from './defaults';
 import {parseConfigFile, writeBackSchemaIfNeeded} from './configfile';
 import * as github from './github';
-
-function combineResults(results: BenchmarkResult[]): BenchmarkResult {
-  const combined: BenchmarkResult = {
-    ...results[0],
-    millis: [],
-  };
-  for (const result of results) {
-    combined.millis.push(...result.millis);
-  }
-  return combined;
-}
+import {manualMode} from './manual';
+import {automaticMode} from './automatic';
 
 const getVersion = (): string =>
     require(path.join(__dirname, '..', 'package.json')).version;
@@ -210,327 +195,6 @@ $ tach http://example.com
       await Promise.all([...allServers].map((server) => server.close()));
     }
   }
-}
-
-type ServerMap = Map<BenchmarkSpec, Server>;
-
-function specUrl(
-    spec: BenchmarkSpec, servers: ServerMap, config: Config): string {
-  if (spec.url.kind === 'remote') {
-    return spec.url.url;
-  }
-  const server = servers.get(spec);
-  if (server === undefined) {
-    throw new Error('Internal error: no server for spec');
-  }
-  if (config.remoteAccessibleHost !== '' &&
-      spec.browser.remoteUrl !== undefined) {
-    return 'http://' + config.remoteAccessibleHost + ':' + server.port +
-        spec.url.urlPath + spec.url.queryString;
-  }
-  return server.url + spec.url.urlPath + spec.url.queryString;
-}
-
-/**
- * Let the user run benchmarks manually. This process will not exit until
- * the user sends a termination signal.
- */
-async function manualMode(config: Config, servers: ServerMap) {
-  if (config.savePath) {
-    throw new Error(`Can't save results in manual mode`);
-  }
-
-  console.log('\nVisit these URLs in any browser:');
-  const allServers = new Set<Server>([...servers.values()]);
-  for (const spec of config.benchmarks) {
-    console.log();
-    if (spec.url.kind === 'local') {
-      console.log(
-          `${spec.name}${spec.url.queryString}` +
-          (spec.url.version !== undefined ? ` [@${spec.url.version.label}]` :
-                                            ''));
-    }
-    console.log(ansi.format(`[yellow]{${specUrl(spec, servers, config)}}`));
-  }
-
-  console.log(`\nResults will appear below:\n`);
-  for (const server of [...allServers]) {
-    (async function() {
-      while (true) {
-        const result = await server.nextResults();
-        server.endSession();
-        console.log(`${result.millis.toFixed(3)} ms`);
-      }
-    })();
-  }
-}
-
-interface Browser {
-  name: string;
-  driver: webdriver.WebDriver;
-  initialTabHandle: string;
-}
-
-async function automaticMode(config: Config, servers: ServerMap):
-    Promise<Array<ResultStatsWithDifferences>|undefined> {
-  let reportGitHubCheckResults;
-  if (config.githubCheck !== undefined) {
-    const {label, appId, installationId, repo, commit} = config.githubCheck;
-
-    // We can directly store our GitHub App private key as a secret Travis
-    // environment variable (as opposed to committing it as a file and
-    // configuring to Travis decrypt it), but we have to be careful with the
-    // spaces and newlines that PEM files have, since Travis does a raw Bash
-    // substitution when it sets the variable.
-    //
-    // Given a PEM file from GitHub, the following command will escape spaces
-    // and newlines so that it can be safely pasted into the Travis UI. The
-    // spaces will get unescaped by Bash, and we'll unescape newlines ourselves.
-    //
-    //     cat <GITHUB_PEM_FILE>.pem \
-    //         | awk '{printf "%s\\\\n", $0}' | sed 's/ /\\ /g'
-    const appPrivateKey =
-        (process.env.GITHUB_APP_PRIVATE_KEY || '').trim().replace(/\\n/g, '\n');
-    if (appPrivateKey === '') {
-      throw new Error(
-          'Missing or empty GITHUB_APP_PRIVATE_KEY environment variable, ' +
-          'which is required when using --github-check.');
-    }
-    const appToken = github.getAppToken(appId, appPrivateKey);
-    const installationToken =
-        await github.getInstallationToken({installationId, appToken});
-
-    // Create the initial Check Run run now, so that it will show up in the
-    // GitHub UI as pending.
-    const checkId =
-        await github.createCheckRun({label, repo, commit, installationToken});
-
-    // We'll call this after we're done to complete the Check Run.
-    reportGitHubCheckResults = async ({fixed, unfixed}: AutomaticResults) => {
-      const markdown = horizontalHtmlResultTable(fixed) + '\n' +
-          verticalHtmlResultTable(unfixed);
-      await github.completeCheckRun(
-          {label, repo, installationToken, checkId, markdown});
-    };
-  }
-
-  console.log('Running benchmarks\n');
-
-  const specs = config.benchmarks;
-  const bar = new ProgressBar('[:bar] :status', {
-    total: specs.length * (config.sampleSize + /** warmup */ 1),
-    width: 58,
-  });
-
-  const browsers = new Map<string, Browser>();
-  for (const {browser} of specs) {
-    const sig = browserSignature(browser);
-    if (browsers.has(sig)) {
-      continue;
-    }
-    bar.tick(0, {status: `launching ${browser.name}`});
-    // It's important that we execute each benchmark iteration in a new tab.
-    // At least in Chrome, each tab corresponds to process which shares some
-    // amount of cached V8 state which can cause significant measurement
-    // effects. There might even be additional interaction effects that
-    // would require an entirely new browser to remove, but experience in
-    // Chrome so far shows that new tabs are neccessary and sufficient.
-    const driver = await makeDriver(browser);
-    const tabs = await driver.getAllWindowHandles();
-    // We'll always launch new tabs from this initial blank tab.
-    const initialTabHandle = tabs[0];
-    browsers.set(sig, {name: browser.name, driver, initialTabHandle});
-  }
-
-  const runSpec = async(spec: BenchmarkSpec): Promise<BenchmarkResult> => {
-    let server;
-    if (spec.url.kind === 'local') {
-      server = servers.get(spec);
-      if (server === undefined) {
-        throw new Error('Internal error: no server for spec');
-      }
-    }
-
-    const url = specUrl(spec, servers, config);
-    const {driver, initialTabHandle} =
-        browsers.get(browserSignature(spec.browser))!;
-
-    let millis;
-    let bytesSent = 0;
-    let userAgent = '';
-    // TODO(aomarks) Make maxAttempts and timeouts configurable.
-    const maxAttempts = 3;
-    for (let attempt = 1;; attempt++) {
-      await openAndSwitchToNewTab(driver, spec.browser);
-      await driver.get(url);
-
-      if (spec.measurement === 'fcp') {
-        millis = await pollForFirstContentfulPaint(driver);
-      } else if (spec.measurement === 'global') {
-        millis = await pollForGlobalResult(driver);
-      } else {  // bench.start() and bench.stop() callback
-        if (server === undefined) {
-          throw new Error('Internal error: no server for spec');
-        }
-        millis = (await server.nextResults()).millis;
-      }
-
-      // Close the active tab (but not the whole browser, since the
-      // initial blank tab is still open).
-      await driver.close();
-      await driver.switchTo().window(initialTabHandle);
-
-      if (server !== undefined) {
-        const session = server.endSession();
-        bytesSent = session.bytesSent;
-        userAgent = session.userAgent;
-      }
-
-      if (millis !== undefined || attempt >= maxAttempts) {
-        break;
-      }
-
-      console.log(
-          `\n\nFailed ${attempt}/${maxAttempts} times ` +
-          `to get a ${spec.measurement} measurement ` +
-          `in ${spec.browser.name} from ${url}. Retrying.`);
-    }
-
-    if (millis === undefined) {
-      console.log();
-      throw new Error(
-          `\n\nFailed ${maxAttempts}/${maxAttempts} times ` +
-          `to get a ${spec.measurement} measurement ` +
-          `in ${spec.browser.name} from ${url}. Aborting.\n`);
-    }
-
-    return {
-      name: spec.name,
-      queryString: spec.url.kind === 'local' ? spec.url.queryString : '',
-      version: spec.url.kind === 'local' && spec.url.version !== undefined ?
-          spec.url.version.label :
-          '',
-      millis: [millis],
-      bytesSent,
-      browser: spec.browser,
-      userAgent,
-    };
-  };
-
-  // Do one throw-away run per benchmark to warm up our server (especially
-  // when expensive bare module resolution is enabled), and the browser.
-  for (let i = 0; i < specs.length; i++) {
-    const spec = specs[i];
-    bar.tick(0, {
-      status: `warmup ${i + 1}/${specs.length} ${benchmarkOneLiner(spec)}`,
-    });
-    await runSpec(spec);
-    bar.tick(1);
-  }
-
-  const specResults = new Map<BenchmarkSpec, BenchmarkResult[]>();
-  for (const spec of specs) {
-    specResults.set(spec, []);
-  }
-
-  // Always collect our minimum number of samples.
-  const numRuns = specs.length * config.sampleSize;
-  let run = 0;
-  for (let sample = 0; sample < config.sampleSize; sample++) {
-    for (const spec of specs) {
-      bar.tick(0, {
-        status: `${++run}/${numRuns} ${benchmarkOneLiner(spec)}`,
-      });
-      specResults.get(spec)!.push(await runSpec(spec));
-      if (bar.curr === bar.total - 1) {
-        // Note if we tick with 0 after we've completed, the status is
-        // rendered on the next line for some reason.
-        bar.tick(1, {status: 'done'});
-      } else {
-        bar.tick(1);
-      }
-    }
-  }
-
-  const makeResults = () => {
-    const results: BenchmarkResult[] = [];
-    for (const sr of specResults.values()) {
-      results.push(combineResults(sr));
-    }
-    const withStats = results.map((result): ResultStats => ({
-                                    result,
-                                    stats: summaryStats(result.millis),
-                                  }));
-    return computeDifferences(withStats);
-  };
-
-  let hitTimeout = false;
-  if (config.timeout > 0) {
-    console.log();
-    const timeoutMs = config.timeout * 60 * 1000;  // minutes -> millis
-    const startMs = Date.now();
-    let run = 0;
-    let sample = 0;
-    let elapsed = 0;
-    while (true) {
-      if (horizonsResolved(makeResults(), config.horizons)) {
-        console.log();
-        break;
-      }
-      if (elapsed >= timeoutMs) {
-        hitTimeout = true;
-        break;
-      }
-      // Run batches of 10 additional samples at a time for more presentable
-      // sample sizes, and to nudge sample sizes up a little.
-      for (let i = 0; i < 10; i++) {
-        sample++;
-        for (const spec of specs) {
-          run++;
-          elapsed = Date.now() - startMs;
-          const remainingSecs =
-              Math.max(0, Math.round((timeoutMs - elapsed) / 1000));
-          const mins = Math.floor(remainingSecs / 60);
-          const secs = remainingSecs % 60;
-          process.stdout.write(
-              `\r${spinner[run % spinner.length]} Auto-sample ${sample} ` +
-              `(timeout in ${mins}m${secs}s)` + ansi.erase.inLine(0));
-          specResults.get(spec)!.push(await runSpec(spec));
-        }
-      }
-    }
-  }
-
-  // Close the browsers by closing each of their last remaining tabs.
-  await Promise.all([...browsers.values()].map(({driver}) => driver.close()));
-
-  const withDifferences = makeResults();
-  console.log();
-  const {fixed, unfixed} = automaticResultTable(withDifferences);
-  console.log(horizontalTermResultTable(fixed));
-  console.log(verticalTermResultTable(unfixed));
-
-  if (hitTimeout === true) {
-    console.log(ansi.format(
-        `[bold red]{NOTE} Hit ${config.timeout} minute auto-sample timeout` +
-        ` trying to resolve horizon(s)`));
-    console.log('Consider a longer --timeout or different --horizon');
-  }
-
-  if (config.savePath) {
-    const session = await makeSession(withDifferences.map((s) => s.result));
-    await fsExtra.writeJSON(config.savePath, session);
-  }
-
-  if (config.csvFile) {
-    await fsExtra.writeFile(config.csvFile, formatCsv(withDifferences));
-  }
-
-  if (reportGitHubCheckResults !== undefined) {
-    await reportGitHubCheckResults({fixed, unfixed});
-  }
-
-  return withDifferences;
 }
 
 /** Parse horizon flags into signed horizon values. */
